pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "customer-support-env"
version = "1.0.0"
description = "OpenEnv-compatible customer support ticket resolution environment"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=15.0.1",
    "pydantic>=2.7.1",
    "PyYAML>=6.0.1",
    "requests>=2.32.3",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
inference = [
    "openai>=2.7.2",
    "python-dotenv>=1.0.1",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["customer_support_env*"]

[tool.setuptools.package-data]
customer_support_env = ["data/*.json"]
